Seperated the folders fro auth and root
