+++ screens/Auth/LoginScreen.js
@@ -4,9 +4,10 @@ import Icon from 'react-native-vector-icons/FontAwesome5';
 import { Card, Text, Input, Button } from 'react-native-elements';
 import firebase from 'firebase';
 import * as GoogleSignIn from 'expo-google-sign-in';
-import Colors from '../constants/Colors';
+import Colors from '../../constants/Colors';
+import { googleConfig } from '../../constants/firebase.config';
 
-class SignupScreen extends Component {
+class LoginScreen extends Component {
     state = {
         loading: false,
         user: null,
@@ -20,17 +21,63 @@ class SignupScreen extends Component {
     componentDidMount() {
     }
 
-    createUserWithEmailAndPassword() {
+    initAsync = async () => {
+        await GoogleSignIn.initAsync({
+            clientId: googleConfig.client_id
+        });
+
+    }
+
+    _syncUserWithStateAsync = async () => {
+        const user = await GoogleSignIn.signInSilentlyAsync();
+        console.log(user);
+        // TODO :: now the use object has been received from google do a firebase signup
+        this.setState({ user: user });
+    }
+
+    signOutAsync = async () => {
+        await GoogleSignIn.signOutAsync();
+        this.setState({ user: null });
+    };
+
+    signInAsync = async () => {
+        try {
+            await GoogleSignIn.askForPlayServicesAsync();
+            const { type, user } = await GoogleSignIn.signInAsync();
+            if (type === 'success') {
+                this._syncUserWithStateAsync();
+            }
+        } catch ({ message }) {
+            alert('login: Error:' + message);
+        }
+    }
+
+    onInvokeGoogleSignin = async () => {
+        if (this.state.user) {
+            return this.signOutAsync();
+        } else {
+            return this.signInAsync();
+        }
+    }
+
+    signInWithGoogle() {
+        this.setState({ loading: true });
+        this.onInvokeGoogleSignin().finally(() => {
+            this.setState({ loading: false });
+        })
+    }
+
+    signInWithEmailAndPassword() {
         const { email, password } = this.state;
-        return firebase.auth().createUserWithEmailAndPassword(email, password)
+        return firebase.auth().signInWithEmailAndPassword(email, password)
             .then(res => {
-                return res;
+                console.log(res);
             })
             .catch(err => {
                 console.log(err);
-                alert(err + ' If you own this email, you should login!!');
-                throw new Error(err);
-            });
+                alert(err + ' May be you should signup first!!');
+            })
+        // alert(`Signing in with email ${email}:${password}`);
     }
 
     render() {
@@ -38,7 +85,7 @@ class SignupScreen extends Component {
             <View style={styles.container}>
 
                 <Card
-                    title='Create an account'
+                    title='Login'
                     containerStyle={styles.card}
                 >
                     {(this.state.loading) ? <ActivityIndicator size='large' /> : null}
@@ -75,36 +122,49 @@ class SignupScreen extends Component {
                         buttonStyle={{ backgroundColor: Colors.primary }}
                         loading={false}
                         raised
-                        title="Sign Up"
-                        onPress={() => this.createUserWithEmailAndPassword()}
+                        title="Sign In"
+                        onPress={() => this.signInWithEmailAndPassword()}
                     >
                     </Button>
-
-                </Card>
-                <View>
-                    <Text style={{ textAlign: 'center', fontSize: 16, padding: 15 }}>Already have account?</Text>
+                    <Text style={{textAlign:'center',fontSize:16,padding:15}}>OR, Don't have account yet?</Text>
                     <Button
                         icon={{
-                            name: "sign-in",
+                            name: "user-plus",
                             type: 'font-awesome',
                             size: 15,
                             color: "white"
                         }}
                         buttonStyle={{ backgroundColor: Colors.primary2 }}
-                        titleStyle={{ color: 'white' }}
+                        titleStyle={{color:'white'}}
                         loading={false}
                         raised
-                        title="Login"
-                        onPress={() => this.props.navigation.navigate('Login')}
+                        title="Sign Up"
+                        onPress={() => this.props.navigation.navigate('Signup')}
                     >
                     </Button>
-                </View>
+                </Card>
+                <Button
+                    icon={{
+                        name: "google-plus",
+                        type: 'font-awesome',
+                        size: 15,
+                        color: "black"
+                    }}
+                    loading={false}
+                    raised
+                    title="Sign in with google"
+                    titleStyle={{ color: 'black' }}
+                    containerStyle={styles.buttonContainer}
+                    buttonStyle={styles.button}
+                    onPress={() => this.signInWithGoogle()}
+                >
+                </Button>
             </View>
         );
     }
 }
 
-export default SignupScreen;
+export default LoginScreen;
 
 const styles = StyleSheet.create({
     container: {
