+++ screens/Auth/SignupScreen.js
@@ -4,10 +4,9 @@ import Icon from 'react-native-vector-icons/FontAwesome5';
 import { Card, Text, Input, Button } from 'react-native-elements';
 import firebase from 'firebase';
 import * as GoogleSignIn from 'expo-google-sign-in';
-import Colors from '../constants/Colors';
-import { googleConfig } from './../constants/firebase.config';
+import Colors from '../../constants/Colors';
 
-class LoginScreen extends Component {
+class SignupScreen extends Component {
     state = {
         loading: false,
         user: null,
@@ -21,63 +20,17 @@ class LoginScreen extends Component {
     componentDidMount() {
     }
 
-    initAsync = async () => {
-        await GoogleSignIn.initAsync({
-            clientId: googleConfig.client_id
-        });
-
-    }
-
-    _syncUserWithStateAsync = async () => {
-        const user = await GoogleSignIn.signInSilentlyAsync();
-        console.log(user);
-        // TODO :: now the use object has been received from google do a firebase signup
-        this.setState({ user: user });
-    }
-
-    signOutAsync = async () => {
-        await GoogleSignIn.signOutAsync();
-        this.setState({ user: null });
-    };
-
-    signInAsync = async () => {
-        try {
-            await GoogleSignIn.askForPlayServicesAsync();
-            const { type, user } = await GoogleSignIn.signInAsync();
-            if (type === 'success') {
-                this._syncUserWithStateAsync();
-            }
-        } catch ({ message }) {
-            alert('login: Error:' + message);
-        }
-    }
-
-    onInvokeGoogleSignin = async () => {
-        if (this.state.user) {
-            return this.signOutAsync();
-        } else {
-            return this.signInAsync();
-        }
-    }
-
-    signInWithGoogle() {
-        this.setState({ loading: true });
-        this.onInvokeGoogleSignin().finally(() => {
-            this.setState({ loading: false });
-        })
-    }
-
-    signInWithEmailAndPassword() {
+    createUserWithEmailAndPassword() {
         const { email, password } = this.state;
-        return firebase.auth().signInWithEmailAndPassword(email, password)
+        return firebase.auth().createUserWithEmailAndPassword(email, password)
             .then(res => {
-                console.log(res);
+                return res;
             })
             .catch(err => {
                 console.log(err);
-                alert(err + ' May be you should signup first!!');
-            })
-        // alert(`Signing in with email ${email}:${password}`);
+                alert(err + ' If you own this email, you should login!!');
+                throw new Error(err);
+            });
     }
 
     render() {
@@ -85,7 +38,7 @@ class LoginScreen extends Component {
             <View style={styles.container}>
 
                 <Card
-                    title='Login'
+                    title='Create an account'
                     containerStyle={styles.card}
                 >
                     {(this.state.loading) ? <ActivityIndicator size='large' /> : null}
@@ -122,49 +75,36 @@ class LoginScreen extends Component {
                         buttonStyle={{ backgroundColor: Colors.primary }}
                         loading={false}
                         raised
-                        title="Sign In"
-                        onPress={() => this.signInWithEmailAndPassword()}
+                        title="Sign Up"
+                        onPress={() => this.createUserWithEmailAndPassword()}
                     >
                     </Button>
-                    <Text style={{textAlign:'center',fontSize:16,padding:15}}>OR, Don't have account yet?</Text>
+
+                </Card>
+                <View>
+                    <Text style={{ textAlign: 'center', fontSize: 16, padding: 15 }}>Already have account?</Text>
                     <Button
                         icon={{
-                            name: "user-plus",
+                            name: "sign-in",
                             type: 'font-awesome',
                             size: 15,
                             color: "white"
                         }}
                         buttonStyle={{ backgroundColor: Colors.primary2 }}
-                        titleStyle={{color:'white'}}
+                        titleStyle={{ color: 'white' }}
                         loading={false}
                         raised
-                        title="Sign Up"
-                        onPress={() => this.props.navigation.navigate('Signup')}
+                        title="Login"
+                        onPress={() => this.props.navigation.navigate('Login')}
                     >
                     </Button>
-                </Card>
-                <Button
-                    icon={{
-                        name: "google-plus",
-                        type: 'font-awesome',
-                        size: 15,
-                        color: "black"
-                    }}
-                    loading={false}
-                    raised
-                    title="Sign in with google"
-                    titleStyle={{ color: 'black' }}
-                    containerStyle={styles.buttonContainer}
-                    buttonStyle={styles.button}
-                    onPress={() => this.signInWithGoogle()}
-                >
-                </Button>
+                </View>
             </View>
         );
     }
 }
 
-export default LoginScreen;
+export default SignupScreen;
 
 const styles = StyleSheet.create({
     container: {
